requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.8
Django==4.2
django-graphql-jwt==0.4.0
filelock==3.14.0
graphene==3.3
graphene-django==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
identify==2.5.36
nodeenv==1.9.1
platformdirs==4.2.2
pre-commit==3.7.1
promise==2.3
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
virtualenv==20.26.2
